requirements.txt
streamlit
pydantic
langchain
langchain-openai
langchain-core
python-dotenv
typing-extensions
google-generativeai
Pillow
reportlab





